safe

+++ src/Heap.cpp
@@ -62,7 +62,7 @@ void Heap::heapify_up(int index) {
 void Heap::showHeapArray() {
     cout << "Heap: \n";
 
-    for (int i = 0; i < this->array.size(); i++)
+    for (size_t i = 0; i < this->array.size(); i++)
         cout << this->array[i].first << "\t";
 
     cout << endl;
